fix: moved `addHandler()` to main
