+++ commands/bitch.go
@@ -9,11 +9,13 @@ import (
 func CommandBitch(dg *discordgo.Session, i *discordgo.InteractionCreate) {
 	var message = ""
 
-	options := i.ApplicationCommandData().Options
+	// options := i.ApplicationCommandData().Options
+	options := ParseUserOptions(dg, i)
+	user := options["user"]
 
-	if options[0] != nil && options[0].UserValue(nil).ID != "246732655373189120" {
-		message = "<@" + options[0].UserValue(nil).ID + "> is a bitch."
-	} else if options[0] != nil {
+	if user != nil && user.UserValue(dg).ID != "246732655373189120" {
+		message = "<@" + user.UserValue(dg).ID + "> is a bitch."
+	} else if user != nil {
 		message = "<@" + i.Member.User.ID + "> nice try, you're a bitch."
 	}
 
@@ -23,6 +25,7 @@ func CommandBitch(dg *discordgo.Session, i *discordgo.InteractionCreate) {
 			Content: message,
 		},
 	})
+	fmt.Println("CommandBitch") // __AUTO_GENERATED_PRINTF__
 
 	if err != nil {
 		fmt.Println("Error with function bitch:")
