+++ commands/utils.go
@@ -1,0 +1,13 @@
+package commands
+
+import "github.com/bwmarrin/discordgo"
+
+// ParseUserOptions parses the user option passed to a command, and returns a map of data options
+func ParseUserOptions(sess *discordgo.Session, i *discordgo.InteractionCreate) map[string]*discordgo.ApplicationCommandInteractionDataOption {
+	options := i.ApplicationCommandData().Options
+	optionMap := make(map[string]*discordgo.ApplicationCommandInteractionDataOption, len(options))
+	for _, opt := range options {
+		optionMap[opt.Name] = opt
+	}
+	return optionMap
+}
